Don't show the next game if it's in the past.
ESPN doesn't update the "next game" on the team data until...sometime the week of the next game. Until I build out getting the team schedule from the endpoint, just display a message that we're waiting on ESPN to get the next game data.

Also adds a link to the team clubhouse on ESPN.