Starting to build out some structure that will support future functionality
Added an auth-context that currently just provides a "current user" and a login function that sets the current user. Also added a "settings" indicator when you're viewing the profile of the logged in user. The current user can only be set while in a development environment, though, so no functionality is added to the production app.

Also moved the remaining `_Component.js` files to `index.js`