Added game headlines to display.
If a game has a "headline" in the notes, it's likely a conference championship game or a bowl game. Let's display that.

I've also introduced a function to convert a long-form version of the conference name that appears in the headlines to a shorter version because, really, who wants to read out "American Athletic Conference Championship Game?"